# backend/app/static/swagger.json
{
    "openapi":"3.0.1",
    "info":{
        "title":"Administrador de inventario de Libros",
        "version":"1.0.0"
    },
    "paths":{
        "/api/register":{
            "post":{
                "summary":"Registra un nuevo usuario",
                "tags":[
                    "User"
                ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses":{
                    "201":{
                        "description":"Usuario creado"
                    },
                    "400":{
                        "description":"Solicitud incorrecta"
                    }
                }
            }
        },
        "/api/login":{
            "post":{
                "summary":"Inicia sesion con un usuario existente",
                "tags":[
                    "User"
                ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/Login"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"Inicio de sesion exitoso",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "properties":{
                                        "access_token":{
                                            "type":"string",
                                            "description":"Token de acceso JWT para el usuario"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401":{
                        "description":"Credenciales invalidas"
                    }
                }
            }
        },
        "/api/orders":{
            "get":{
                "summary":"Obtiente la lista de pedidos",
                "tags":[
                    "Admin"
                ],
                "security":[
                    {
                        "JWTAuth":[]
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Lista de pedidos",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref":"#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "summary":"Realiza un pedido",
                "tags":[
                    "Customer"
                ],
                "security":[
                    {
                        "JWTAuth":[]
                    }
                ],
        
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses":{
                    "201":{
                        "description":"Pedido realizado",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref":"#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }

            }
        },
        "/api/orders/{id}":{
            "delete":{
                "summary":"Elimina un pedido",
                "tags":[
                    "Admin"
                ],
                "security":[
                    {
                        "JWTAuth":[]
                    }
                ],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                          "type": "integer"
                        }
                    }
                ],
                "responses":{
                    "204":"Pedido Eliminado"
                }
            }
        },

        "/api/books":{
            "get":{
                "summary":"Obtiene la lista de todos los libros",
                "tags":[
                    "User"
                ],
                
                "responses":{
                    "200":{
                        "description":"Lista de libros",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"array",
                                    "items":{
                                        "$ref":"#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "summary":"Crea un nuevo libro",
                "tags":[
                    "Admin"
                ],
                "security":[
                    {
                        "JWTAuth":[]
                    }
                ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses":{
                    "201":{
                        "description":"Libro creado",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref":"#/components/schemas/Book"   
                                }
                            }
                        }
                    }
                }
            }
            
        },
        "/api/books/{id}":{
            "get":{
                "summary":"Obtiene un libro especifico por su ID",
                "tags":[
                    "User"
                ],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Detalles del libro",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref":"#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404":{
                        "description":"Libro no encontrado"
                    }
                }
            },
            "put":{
                "summary":"Actualiza un libro por us ID",
                "tags":[
                    "Admin"
                ],
                "security":[
                    {
                        "JWTAuth":[]
                    }
                ],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                      "application/json": {
                        "schema": {
                          "$ref": "#/components/schemas/Book"
                        }
                      }
                    }
                },
                "responses":{
                    "200":{
                        "description":"Libro actualizado",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref":"#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404":{
                        "description":"Libro no encontrado"
                    }
                }
            },
            "delete":{
                "summary":"Elimina un libro por su ID",
                "tags":[
                    "Admin"
                ],
                "security":[
                    {
                        "JWTAuth":[]
                    }
                ],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "required":true,
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "responses":{
                    "204":{
                        "description":"Libro Eliminado"
                    },
                    "404":{
                        "description":"Libro no encontrado"
                    }
                }
            }
        }

    },
    "components":{
        "securitySchemes":{
            "JWTAuth":{
                "type":"http",
                "scheme":"bearer",
                "bearerFormat":"JWT"
            }
        },
        "schemas":{
            "User":{
                "tpye":"object",
                "properties":{
                    "id":{
                        "type":"integer",
                        "format":"int32",
                        "readOnly":true
                    },
                    "username":{
                        "type":"string"
                    },
                    "email":{
                        "type":"string"
                    },
                    "password":{
                        "type":"string"
                    },
                    "role":{
                        "type":"array",
                        "items":{
                            "type":"string",
                            "enum":[
                                "admin",
                                "customer"
                            ]
                        }
                    }
                },
                "required":[
                    "username",
                    "email",
                    "password",
                    "role"
                ]
            },
            "Login":{
                "type":"object",
                "properties":{
                    "email":{
                        "type":"string"
                    },
                    "password":{
                        "type":"string"
                    }
                },
                "required":[
                    "email",
                    "password"
                ]
            },
            "Book":{
                "type":"object",
                "properties":{
                    "id":{
                        "type":"integer",
                        "format":"int32",
                        "readOnly":true
                    },
                    "title":{
                        "type":"string"
                    },
                    "price":{
                        "type":"number",
                        "format":"float"
                    },
                    "author":{
                        "type":"string"
                    },
                    "editorial":{
                        "type":"string"
                    },
                    "number_pages":{
                        "type":"integer",
                        "format":"int32"
                    },
                    "copies":{
                        "type":"integer",
                        "format":"int32"
                    },
                    "bookbinding":{
                        "type":"string"
                    },
                    "description":{
                        "type":"string"
                    },
                    "cover":{
                        "type":"string"
                    }
                },
                "required":[
                    "title",
                    "price",
                    "author",
                    "editorial",
                    "number_pages",
                    "copies",
                    "bookbinding",
                    "description",
                    "cover"
                ]
            },
            "Order":{
                "type":"object",
                "properties":{
                    "id":{
                        "type":"integer",
                        "format":"int32",
                        "readOnly":true
                    },
                    "user_id":{
                        "type":"integer"
                        
                    },
                    "title":{
                        "type":"string"
                    },
                    "author":{
                        "type":"string"
                    },
                    "order_date":{
                        "type":"string",
                        "format":"date-time"
                    }

                },
                "required":[
                    "title",
                    "author"
                ]
            }
        }
    },
    "tags":[
        {
            "name":"Admin",
            "description":"Operaciones para administradores"
        },
        {
            "name":"Customer",
            "description":"Operaciones para clientes"
        },
        {
            "name":"User",
            "description":"Operaciones para autentificacion"
        }

    ]

}